fix(field): confirm and correct use the capture; refresh after correcting (RC-1C-HOTFIX)
Confirm: "Confirmar localizacao" opens "Buscando uma localizacao
precisa..." with the live accuracy and Cancel; only an acquired fix is
measured against the point, and the check keeps that exact fix and the
location version it was measured against - the position shown is the
position sent (raw accuracy, not rounded). The no-GPS variant of the
confirm dialog is gone: failures stay in the capture, with the reason and
"Tentar novamente", which opens a new capture.

Correct: with "Usar minha localizacao atual" on, Save captures over the
sheet; failure or cancel keeps the sheet open with the typed fields. The
controller refuses a GPS correction without an in-contract fix instead of
silently downgrading it to address-only. Address-only needs no GPS.

The one-shot, invisible 15 s GPS read in correct is gone; after a
successful correction the bundle is reloaded from the server, now covered
by controller and widget tests (MISSING -> "Confirmada").

Check-in and time clock keep LocationService: there the coordinate is
informational and never blocks.

+++ apps/field/lib/features/execution/data/execution_repository.dart
@@ -98,7 +98,8 @@ class ExecutionRepository {
     required String idempotencyKey,
     double? observedLatitude,
     double? observedLongitude,
-    int? observedAccuracyMeters,
+    // O valor REAL (RC-1C-HOTFIX): o servidor julga a precisão sem arredondar.
+    double? observedAccuracyMeters,
   }) async {
     await _api.post(
       '/service-orders/$orderId/location/confirm',
@@ -127,7 +128,7 @@ class ExecutionRepository {
     String? note,
     double? latitude,
     double? longitude,
-    int? accuracyMeters,
+    double? accuracyMeters,
     String? source,
     Map<String, String?>? address,
   }) async {
